# Conflicts:
#	src/main/java/th/co/ais/mimo/debt/exempt/entity/DccCriteriaHistory.java
#	src/main/java/th/co/ais/mimo/debt/exempt/entity/DccCriteriaHistoryId.java
#	src/main/java/th/co/ais/mimo/debt/exempt/entity/DccCriteriaMasterId.java
#	src/main/java/th/co/ais/mimo/debt/exempt/model/InsertAssignIdReq.java
#	src/main/java/th/co/ais/mimo/debt/exempt/service/DMSEM004CriteriaMasterService.java
#	src/main/java/th/co/ais/mimo/debt/exempt/service/impl/DMSEM004CriteriaMasterServiceImpl.java

+++ src/main/java/th/co/ais/mimo/debt/exempt/entity/DccExemptHistory.java
@@ -40,10 +40,10 @@ public class DccExemptHistory {
     private String cateCode;
 
     @Column(name = "ADD_LOCATION")
-    private Long addLocation;
+    private Integer addLocation;
 
     @Column(name = "UPDATE_LOCATION")
-    private Long updateLocation;
+    private Integer updateLocation;
 
     @Column(name = "ADD_REASON")
     private String addReason;
